requirements.txt
# Core PyTorch ecosystem
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Computer vision and image processing
Pillow>=8.0.0
opencv-python>=4.5.0

# Deep learning and model libraries
timm>=0.6.0
timm[all]>=0.6.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Progress bars and utilities
tqdm>=4.62.0

# Configuration and YAML
PyYAML>=6.0

# Experiment tracking and logging
wandb>=0.12.0
